pyproject.toml
[tool.poetry]
name = "ninja-telegram-bot"
version = "0.1.0"
description = "A bot for creating, saving and executing custom Python users scripts in a Telegram conversation"
authors = ["Sergey Dmitriev <51058739+0niel@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.37"
envparse = "^0.2.0"
pydantic = "^1.9.1"
requests = "^2.28.0"
python-telegram-bot = "^13.12"
psycopg2-binary = "^2.9.3"
pytz = "^2022.1"
python-telegram-bot-pagination = "^0.0.2"
beautifulsoup4 = "^4.11.1"
Pillow = "^9.1.1"
gTTS = "^2.2.4"

[tool.poetry.dev-dependencies]
mypy = "^0.942"
black = "^22.3.0"
isort = "^5.10.1"
autoflake = "^1.4"
flake8 = "^4.0.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"